+++ cargo-jvm/Changelog.md
@@ -1,7 +1,10 @@
-# 0.1.1
+# 0.1.1 - WIP
+
+(still accumlating changes, not yet released)
 
 - Added support for `cdylib` crates in `cargo jvm package --lib`
+- `cargo jvm doctor` emits more info about specific `cargo-jvm` and backend versions/git commits.
 
-# 0.1.0
+# 0.1.0 - current stable
 
 - Initial release of `cargo-jvm`